fixing stripe GW

+++ api-server/models/booking.go
@@ -101,7 +101,7 @@ type BookingPaymentInfo struct {
 	BookingStatusID     zero.Int    `db:"booking_status_id"` // The status of the booking payment (Not_paid, Full_amountPaid, Partial_amount_Paid)
 	BookingStatus       zero.String `json:"booking_status" db:"booking_status"`
 	StripeSessionID     zero.String `db:"stripe_session_id"`
-	AmountPaid          zero.Float  `db:"amount_paid"` // Percentage of the total booking cost paid
+	AmountPaid          zero.Int    `db:"amount_paid"` // Percentage of the total booking cost paid
 	TripName            zero.String `db:"trip_name"`
 	Description         zero.String `db:"description"`
 	MaxParticipants     zero.Int    `json:"max_participants" db:"max_participants"` // Maximum number of people allowed on the trip
